feat: Implement useTextSplitterNode composable for enhanced text splitting functionality in TextSplitterNode component
