+++ frontend/src/components/TextSplitterNode.vue
@@ -21,10 +21,9 @@
 </template>
 
 <script setup>
-import { Handle, useVueFlow } from '@vue-flow/core';
-import { watch, ref, computed, onMounted } from 'vue';
-
-const { getEdges, findNode, onConnect, onConnectStart, onConnectEnd, addEdges } = useVueFlow();
+import { Handle } from '@vue-flow/core';
+import { onMounted } from 'vue';
+import { useTextSplitterNode } from '../composables/useTextSplitterNode';
 
 const props = defineProps({
     id: {
@@ -53,140 +52,16 @@ const props = defineProps({
     },
 });
 
-const endpoint = ref(props.data.inputs?.endpoint || 'http://localhost:8080/api/split-text');
-const text = ref(props.data.inputs?.text || '');
-const outputConnectionCount = ref(0); // Initialize output connection count
+const emit = defineEmits(['update:data']);
 
-watch(
-    () => props.data,
-    (newData) => {
-        endpoint.value = newData.inputs?.endpoint || 'http://localhost:8080/api/split-text';
-        text.value = newData.inputs?.text || '';
-        emit('update:data', { id: props.id, data: newData });
-    },
-    { deep: true }
-);
+// Use the composable and destructure the returned values
+const { endpoint, text, updateNodeData, run } = useTextSplitterNode(props, emit);
 
 onMounted(() => {
     if (!props.data.run) {
-        props.data.run = run
+        props.data.run = run;
     }
-})
-
-// Watch for changes in edges and update output connections accordingly
-watch(
-    () => getEdges.value,
-    () => {
-        updateOutputConnections();
-    },
-    { deep: true }
-);
-
-const updateOutputConnections = async () => {
-    // Find connected output edges
-    const outputEdges = getEdges.value.filter((edge) => edge.source === props.id);
-    outputConnectionCount.value = outputEdges.length;
-
-    // Update the node data with the new output connection count
-    updateNodeData();
-};
-
-const updateNodeData = async () => {
-    const updatedData = {
-        ...props.data,
-        inputs: {
-            endpoint: endpoint.value,
-            text: text.value,
-        },
-        outputs: {}, // Initialize outputs as an empty object
-        num_chunks: outputConnectionCount.value
-    };
-
-    // Get connected input edges and their source nodes
-    const inputEdges = getEdges.value.filter((edge) => edge.target === props.id && edge.targetHandle === 'input');
-    for (const edge of inputEdges) {
-        const sourceNode = findNode(edge.source);
-        if (sourceNode && sourceNode.data.outputs) {
-            // Assuming the text to be split is in the 'output' property of the source node's outputs
-            if (sourceNode.data.outputs[edge.sourceHandle]) {
-                updatedData.inputs.text = sourceNode.data.outputs[edge.sourceHandle];
-            }
-        }
-    }
-    emit('update:data', { id: props.id, data: updatedData });
-};
-
-async function run() {
-    console.log('Running TextSplitterNode:', props.id);
-
-    const connectedTargetEdges = getEdges.value.filter(
-        (edge) => edge.target === props.id
-    );
-
-    // Get the first connected edge
-    const targetEdge = connectedTargetEdges[0];
-
-    console.log('Connected target edge:', targetEdge);
-
-    // Get the source node of the connected edge
-    const sourceNode = findNode(targetEdge.source);
-
-    console.log('Source node:', sourceNode);
-
-    // Get the response value from the source node's outputs
-    const response = sourceNode.data.outputs.result.output;
-
-    console.log('Response:', response);
-
-    // Update the input text with the response value
-    text.value = response;
-
-    // Update the node data with the new input text
-    updateNodeData();
-
-    // Get the source edges
-    const sourceEdges = getEdges.value.filter(
-        (edge) => edge.source === props.id
-    );
-
-    // Get the target nodes of the source edges
-    const targetNodes = sourceEdges.map((edge) => findNode(edge.target));
-
-    console.log('Target nodes:', targetNodes);
-
-    // Split the text by the number of target nodes, for example if there are two target nodes, then the text should be split into two strings:
-    // "This is my example test" -> ["This is my", "example test"]
-    
-    // First get a count of the source connections
-    const sourceCount = sourceEdges.length;
-
-    // Split the text into chunks based on the number of source connections
-    const words = text.value.split(' ');
-    const wordsPerChunk = Math.ceil(words.length / sourceCount);
-    const chunks = [];
-
-    for (let i = 0; i < sourceCount; i++) {
-        const start = i * wordsPerChunk;
-        const end = Math.min(start + wordsPerChunk, words.length);
-        chunks.push(words.slice(start, end).join(' '));
-
-        // Update the outputs for each target node
-        if (targetNodes[i]) {
-            console.log('Updating target node:', targetNodes[i]);
-            targetNodes[i].data.inputs.response += chunks[i];
-            targetNodes[i].data.inputs.text += chunks[i];
-
-            updateNodeData();
-        }
-    }
-
-
-    console.log('Chunks:', chunks);
-
-}
-
-const emit = defineEmits(['update:data']);
-
+});
 </script>
 
 <style scoped>
